Encapsulate legend data code
