+++ client/plots/boxplot/view/View.ts
@@ -3,7 +3,8 @@ import { ScaleLinear, scaleLinear } from 'd3-scale'
 import { axisstyle } from '#src/client'
 import { axisTop } from 'd3-axis'
 import type { BoxPlotDom, BoxPlotSettings } from '../BoxPlot'
-import type { ViewData, LegendItemEntry } from '../viewModel/ViewModel'
+import type { LegendItemEntry } from '../viewModel/LegendData'
+import type { ViewData } from '../viewModel/ViewModel'
 import type { MassAppApi } from '#mass/types/mass'
 import { BoxPlotToolTips } from './BoxPlotToolTips'
 import { BoxPlotLabelMenu } from './BoxPlotLabelMenu'
